project/requirements.txt
flask
flask-cors
google-generativeai
gunicorn  
openai
